==> Drug_Repositioning/5_new_drug/1_sticth_to_drugbank_id.ipynb <==
import csv
import numpy as np
import pandas as pd
import time
from urllib2 import urlopen # python2
import urllib2
#import urllib.request #python3
#from urllib.request import urlopen
#import urllib.parse


raw_sider = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/meddra_all_label_se_header.tsv")

stitch_cid = pd.read_table("/home/share/aschoi/data/COCONUT_Herbal/7_Ezyme_tool/STITCH_CID.tsv")

print(len(raw_sider))

raw_sider_pt = raw_sider[raw_sider.MedDRA_concept_type == 'PT']
print(len(raw_sider_pt))
raw_sider_pt.head(1)

print(len(raw_sider_pt.drop_duplicates(subset = ['STITCH_compound_id_flat','STITCH_compound_id_stereo', 'UMLS_concept_id_for_MedDRA', 'side_effect_name'])))

sider = raw_sider_pt[['STITCH_compound_id_flat','STITCH_compound_id_stereo', 'UMLS_concept_id_for_MedDRA', 'side_effect_name']]

print(len(sider))

sider_wo_dup = sider.drop_duplicates()
print(len(sider_wo_dup))
sider_wo_dup.head()

len(sider_wo_dup.drop_duplicates(subset='side_effect_name'))

stitch_cid.head(1)

sider_wo_dup2 = sider_wo_dup.drop_duplicates(subset=['STITCH_compound_id_flat', 'STITCH_compound_id_stereo'])

print(len(sider_wo_dup2))
sider_wo_dup2.head()

stitch_cid[(stitch_cid.flat_chemical == 'CID100216416') & (stitch_cid.stereo_chemical == 'CID000216416')]

sider_wo_dup2[(sider_wo_dup2.STITCH_compound_id_flat=='CID100216416') & (sider_wo_dup2.STITCH_compound_id_stereo == 'CID000216416')]

query_pair = list()
for index, row in sider_wo_dup2.iterrows():
    record =row.tolist()
    
    #found = stitch_cid.query('@record[0] == flat_chemical and @record[1]==stereo_chemical')
    found = stitch_cid[(stitch_cid.flat_chemical == record[0]) & (stitch_cid.stereo_chemical == record[1])]
    query_pc_id = found['source_id'].tolist()
    
    query_pair.append([record[0], record[1], query_pc_id])

df_query_pair = pd.DataFrame()
for i in query_pair:
    for j in i[2]:
        temp_df = pd.DataFrame([[i[0], i[1], j]], columns = ['stitch_flat', 'stitch_stereo', 'pubchem_cid'])
        df_query_pair = df_query_pair.append(temp_df)
df_query_pair.to_csv("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/1_sider_pubchem.tsv", sep='\t', index=False)  
print(len(df_query_pair))

df_query_pair = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/1_sider_pubchem.tsv")

print(len(df_query_pair))
df_query_pair.head()

# using UniChem REST API, convert pubchemID to DrugBank ID
with open("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/3_SIDER_DrugBank_id.tsv", "wa") as resultFile:
    features = ['STITCH_compound_id_flat', 'STITCH_compound_id_stereo', 'PubChem_ID', 'DrugBank_ID']
    resultTsv = csv.DictWriter(resultFile, features, dialect=csv.excel_tab)
    resultTsv.writeheader()
    
    #cnt = 0
    url_base = 'https://www.ebi.ac.uk/unichem/rest/src_compound_id/'
    print'%d : ' % cnt + time.strftime('%a %H:%M:%S')
    for index, row in df_query_pair[cnt-1:].iterrows():
        record = row.tolist()
        stitch_id_flat = record[0]
        stitch_id_stereo = record[1]
        pubchem_id = record[2]
        url = url_base + str(pubchem_id) + '/22/2'

        try:
            f = urlopen(url)
        except urllib2.HTTPError, e:
            if e.code == 404:
                cnt = cnt + 1
                if cnt % 1000 == 0 :
                    print '%d : '% cnt + time.strftime('%a %H:%M:%S')
                continue
            else:
                print 'e.code : %d, url : %s, cnt : d' % (e.code, url, cnt)
                raise
        except urllib2.URLError, e:
            if e.args[0][0] == 110: # connection time out
                #sleep_secs = attempt ** 2
                print e
                print url
                print '  cnt = %d, error time : ' % (cnt) + time.strftime('%a %H:%M:%S')
                print '    Retrying in %d seconds...' % sleep_secs
                time.sleep(sleep_secs)
                cnt= cnt +1
                if cnt % 1000 == 0 :
                    print '%d : '% cnt + time.strftime('%a %H:%M:%S')
                continue 
            else:
                print e
                print '   cnt : %d, error time : ' % (cnt) + time.strftime('%a %H:%M:%S')
                raise
        else:          
            all_data = f.readlines()
            if all_data[0] == '[]':
                cnt = cnt + 1
            else:
                #db_id=all_data[0].decode("utf-8").split(':')[1]  
                db_id=all_data[0] 
                resultTsv.writerow({'STITCH_compound_id_flat':stitch_id_flat, 'STITCH_compound_id_stereo':stitch_id_stereo, 
                                'PubChem_ID' : pubchem_id, 'DrugBank_ID' : db_id})     
                cnt = cnt + 1
            if cnt % 1000 == 0 :
                print '%d : '% cnt + time.strftime('%a %H:%M:%S')

print time.strftime('%a %H:%M:%S')


df_query_pair.head(1)

print(len(df_query_pair))
print(len(df_query_pair.drop_duplicates(subset=['stitch_flat', 'stitch_stereo'])))

df_raw = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/3_SIDER_DrugBank_id.tsv")

print(len(df_raw))
print(len(df_raw.drop_duplicates()))
print(len(df_raw.drop_duplicates(subset=['STITCH_compound_id_flat', 'STITCH_compound_id_stereo'])))
df_raw.head()

result = pd.DataFrame()
for index, row in df_raw.iterrows():
    record = row.tolist()
    db_ids = record[3].split(',')
    for i in db_ids:
        db_id = i.split(':')[1].lstrip('\"').rstrip('\"}]')
        temp_df = pd.DataFrame([[record[0], record[1], record[2], db_id]], columns = ['STITCH_compound_id_flat',
                                                                                      'STITCH_compound_id_stereo', 
                                                                                      'PubChem_ID', 'DrugBank_ID'])
        result = result.append(temp_df)
result.index = range(0, len(result))
    
        

print len(result)
print len(result.drop_duplicates())
print len(result.drop_duplicates(subset = ['STITCH_compound_id_flat','STITCH_compound_id_stereo', 'DrugBank_ID']))
print len(result.drop_duplicates(subset = ['STITCH_compound_id_flat','STITCH_compound_id_stereo']))
print len(result.drop_duplicates(subset = ['DrugBank_ID']))

result.to_csv("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/3_2_SIDER_DrugBank_id.tsv", sep='\t', index=False)

result.head()

result[result.duplicated(subset = ['STITCH_compound_id_flat','STITCH_compound_id_stereo', 'DrugBank_ID'], keep=False)].sort_values(by = 'PubChem_ID')#.sort_values(by = 'DrugBank_ID')

stitch_db_table = result[['STITCH_compound_id_flat','STITCH_compound_id_stereo', 'DrugBank_ID']].drop_duplicates()

print len(stitch_db_table)
print len(result[['STITCH_compound_id_flat','STITCH_compound_id_stereo', 'DrugBank_ID']])

stitch_db_table.to_csv("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/3_3_SIDER_STITCHid_DrugBankid.tsv", sep='\t', index=False)

stitch_db_table[stitch_db_table.duplicated(subset=['STITCH_compound_id_flat','STITCH_compound_id_stereo'], keep=False)]
