==> Drug_Repositioning/5_new_drug/2_offside_to_drugbank_id.ipynb <==
import csv
import numpy as np
import pandas as pd
import time
from urllib2 import urlopen # python2
import urllib2

raw_offisde = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/3003377s-offsides.tsv")
stitch_cid = pd.read_table("/home/share/aschoi/data/COCONUT_Herbal/7_Ezyme_tool/STITCH_CID.tsv")

print len(raw_offisde)
raw_offisde.head()

print(len(raw_offisde))
offside = raw_offisde[['stitch_id','drug', 'umls_id', 'event']]
offside.head()

offside_wo_dup = offside.drop_duplicates(subset=['stitch_id', 'drug'])
print(len(offside_wo_dup))
offside_wo_dup.head()

print len(offside.drop_duplicates(subset='umls_id'))
print len(offside.drop_duplicates(subset='event'))

stitch_cid.head(1)

# using UniChem REST API, convert pubchemID to DrugBank ID
with open("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/4_OFFSIDE_DrugBank_id.tsv", "w") as resultFile:
    features = ['STITCH_compound_id_stereo', 'PubChem_ID', 'DrugBank_ID']
    resultTsv = csv.DictWriter(resultFile, features, dialect=csv.excel_tab)
    resultTsv.writeheader()
    
    cnt = 0
    result = list()
    url_base = 'https://www.ebi.ac.uk/unichem/rest/src_compound_id/'
    print '%d : '% cnt + time.strftime('%a %H:%M:%S')
    for index, row in offside_wo_dup[cnt:].iterrows():
        record = row.tolist()
        found = stitch_cid[stitch_cid.stereo_chemical == record[0]]
        pubchem_ids = found['source_id']
        for j in pubchem_ids:
            url = url_base + str(j) + '/22/2'
            try:
                f=urlopen(url)
            except urllib2.HTTPError, e:
                if e.code == 404:
                    cnt = cnt + 1
                    if cnt % 1000 == 0 :
                        print '%d : '% cnt + time.strftime('%a %H:%M:%S')
                    continue
                else:
                    print 'e.code : %d, url : %s, cnt : d' % (e.code, url, cnt)
                    raise
            except urllib2.URLError, e:
                if e.args[0][0] == 110: # connection time out
                    sleep_secs = attempt ** 2
                    print e
                    print url
                    print '  cnt = %d, error time : ' % (cnt) + time.strftime('%a %H:%M:%S')
                    print '    Retrying in %d seconds...' % sleep_secs
                    time.sleep(sleep_secs)
                    cnt= cnt +1
                    if cnt % 1000 == 0 :
                        print '%d : '% cnt + time.strftime('%a %H:%M:%S')
                    continue
                else:
                    print e
                    print '   cnt : %d, error time : ' % (cnt) + time.strftime('%a %H:%M:%S')
                    raise
        
            else:
                all_data = f.readlines()
                if all_data[0] == '[]':
                    cnt = cnt + 1
                else:
                    db_id=all_data[0]
                    #result.append([record[0], record[1], db_id])
                    resultTsv.writerow({'STITCH_compound_id_stereo':record[0],'PubChem_ID' : j, 'DrugBank_ID' : db_id})
                    cnt = cnt + 1

                if cnt % 1000 == 0:
                    print '%d : '% cnt + time.strftime('%a %H:%M:%S')
print time.strftime('%a %H:%M:%S')

cnt

raw_db = pd.read_table('/home/share/aschoi/data/Drug_Repositioning/5_new_drug/4_1_raw_OFFSIDE_DrugBank_id.tsv')

raw_db.head()

stitch_cid[stitch_cid.stereo_chemical == 'CID000000119']

offside_wo_dup[offside_wo_dup.stitch_id == 'CID000000119']

print len(raw_db)
print len(raw_db.drop_duplicates(subset = 'STITCH_compound_id_stereo'))

se_db_id = raw_db['DrugBank_ID']

raw_db.head(1)

result = pd.DataFrame()
for index, row in raw_db.iterrows():
    record = row.tolist()
    db_ids = record[2].split(',')
    for i in db_ids:
        db_id = i.split(':')[1].lstrip('\"').rstrip('\"}]')
        temp_df = pd.DataFrame([[record[0], record[1], db_id]], columns = ['STITCH_compound_id_stereo', 'PubChem_ID', 'DrugBank_ID'])
        result = result.append(temp_df)
result.index = range(0, len(result))
    
        

print len(result)
print len(result.drop_duplicates(subset = ['STITCH_compound_id_stereo', 'DrugBank_ID']))
print len(result.drop_duplicates(subset = ['STITCH_compound_id_stereo']))
print len(result.drop_duplicates(subset = ['DrugBank_ID']))

result.to_csv("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/4_2_OFFSIDE_DrugBank_id.tsv", sep='\t', index=False)

result.head()

result[result.duplicated(subset = ['STITCH_compound_id_stereo', 'DrugBank_ID'], keep=False)].sort_values(by = 'DrugBank_ID')
#result[result.duplicated(subset = ['DrugBank_ID'], keep=False)].sort_values(by = 'DrugBank_ID')

stitch_db_table = result[['STITCH_compound_id_stereo', 'DrugBank_ID']].drop_duplicates()

len(stitch_db_table)

stitch_db_table.to_csv("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/4_3_OFFSIDE_STITCHid_DrugBankid.tsv", sep='\t', index=False)

stitch_db_table[stitch_db_table.duplicated(subset='STITCH_compound_id_stereo', keep=False)]

result[result.STITCH_compound_id_stereo =='CID000000753']

result[result.STITCH_compound_id_stereo =='CID000002764']

result[result.STITCH_compound_id_stereo =='CID000005578']

result[result.STITCH_compound_id_stereo =='CID000193962']
