==> Drug_Repositioning/5_new_drug/3_sider_offside_drugbankid_compare_association.ipynb <==
import csv
import numpy as np
import pandas as pd
import time
from urllib2 import urlopen # python2
import urllib2

sider_id = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/3_3_SIDER_STITCHid_DrugBankid.tsv")
sider_se = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/3_4_SIDER_db_side_effect.tsv")
offside_id = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/4_3_OFFSIDE_STITCHid_DrugBankid.tsv")
offside_se = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/4_4_OFFSIDE_db_side_effect.tsv")

df_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/3_PREDICT_id_association.tsv")

print(len(df_asso))
print(len(df_asso.drop_duplicates(subset='DrugBank_ID')))
print(len(df_asso.drop_duplicates(subset='OMIM_ID')))
df_asso.head(2)

sider_merged = pd.merge(left=sider_id, right=sider_se)

sider_db_se = sider_merged[['DrugBank_ID','UMLS_concept_id_for_MedDRA','side_effect_name']]
print len(sider_db_se)
print len(sider_db_se.drop_duplicates())

sider_db_se.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/3_5_SIDER_DrugBank_sideffect.tsv", sep='\t', index=False)
sider_db_se.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/3_5_SIDER_DrugBank_sideffect.tsv", sep='\t', index=False)

off_merged = pd.merge(left=offside_id, right=offside_se, left_on = 'STITCH_compound_id_stereo', right_on='stitch_id')

off_db_se = off_merged[['DrugBank_ID', 'umls_id', 'event']]
print len(off_db_se)
print len(off_db_se.drop_duplicates())

off_db_se.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/4_5_OFFSIDE_DrugBank_sideffect.tsv", sep='\t', index=False)
off_db_se.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/4_5_OFFSIDE_DrugBank_sideffect.tsv", sep='\t', index=False)

sider_db_se = sider_db_se.rename(columns={'UMLS_concept_id_for_MedDRA':'umls_id', 'side_effect_name':'event'})

concated = pd.concat([sider_db_se, off_db_se])
print len(concated)

drugbank_se = concated.drop_duplicates(['DrugBank_ID', 'umls_id'])

print len(drugbank_se)
print len(drugbank_se.drop_duplicates('DrugBank_ID'))
print len(drugbank_se.drop_duplicates('umls_id'))
print len(drugbank_se.drop_duplicates('event'))

drugbank_se.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/5_DrugBank_SideEffect_total.tsv", sep='\t', index=False)
drugbank_se.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/5_DrugBank_SideEffect_total.tsv", sep='\t', index=False)

print len(df_asso)
df_asso_se = df_asso[df_asso['DrugBank_ID'].isin(drugbank_se['DrugBank_ID'])]
print len(df_asso_se)
print len(df_asso_se.drop_duplicates('DrugBank_ID'))
print len(df_asso_se.drop_duplicates('OMIM_ID'))

gold_se = drugbank_se[drugbank_se['DrugBank_ID'].isin(df_asso['DrugBank_ID'])]
print len(gold_se)
print len(gold_se.drop_duplicates())
print len(gold_se.drop_duplicates('DrugBank_ID'))
print len(gold_se.drop_duplicates('umls_id'))

gold_se.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/PREDICT_data/1_gold_drug_se.tsv", sep='\t', index=False)
gold_se.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/PREDICT_data/1_gold_drug_se.tsv", sep='\t', index=False)

len(df_sider.drop_duplicates(subset=['STITCH_compound_id_flat','STITCH_compound_id_stereo']))

print(len(df_offside))
df_offside.head()

print len(df_asso[df_asso['DrugBank_ID'].isin(df_sider['DrugBank_ID'])].drop_duplicates(subset='DrugBank_ID'))
df_asso[df_asso['DrugBank_ID'].isin(df_sider['DrugBank_ID'])].drop_duplicates(subset='DrugBank_ID').head()

print len(df_asso[df_asso['DrugBank_ID'].isin(df_offside['DrugBank_ID'])].drop_duplicates(subset='DrugBank_ID'))
df_asso[df_asso['DrugBank_ID'].isin(df_offside['DrugBank_ID'])].drop_duplicates(subset='DrugBank_ID').head()

a = pd.concat([df_sider['DrugBank_ID'], df_offside['DrugBank_ID']])

b = a.drop_duplicates()

print len(df_sider['DrugBank_ID'].drop_duplicates())
print len(df_offside['DrugBank_ID'].drop_duplicates())
print len(a)
print len(b)

print len(df_asso['DrugBank_ID'].drop_duplicates())
print len(df_asso[df_asso['DrugBank_ID'].isin(b)].drop_duplicates(subset='DrugBank_ID'))
print len(df_asso)
print len(df_asso[df_asso['DrugBank_ID'].isin(b)])
df_asso[df_asso['DrugBank_ID'].isin(b)].drop_duplicates(subset='DrugBank_ID').head()

c = df_asso[df_asso['DrugBank_ID'].isin(b)]
print(len(df_asso.drop_duplicates(subset='OMIM_ID')))
print(len(c.drop_duplicates(subset='OMIM_ID')))
