==> Drug_Repositioning/5_new_drug/4_sider_offside_drugbankid_compare_association(CDataset).ipynb <==
import csv
import numpy as np
import pandas as pd
import time
from urllib2 import urlopen # python2
import urllib2
from itertools import product

di = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/7_new_association/MBiRW-master/Datasets/CDatasets/DiseasesName", header=None)
dr = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/7_new_association/MBiRW-master/Datasets/CDatasets/DrugsName", header=None)
didr = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/7_new_association/MBiRW-master/Datasets/CDatasets/DiDrAMat", header=None, sep='   ')

print(len(di))
di = di.rename(columns={0:'disease_id'})
di.head(1)

print(len(dr))
dr = dr.rename(columns={0:'drug_id'})
dr.head(1)

drug_id = dr['drug_id'].tolist()
didr.columns = drug_id
disease_id = di['disease_id'].tolist()
didr.index = disease_id
didr.head(1)

didr[drug_id[0]].ix[disease_id[disease_id.index('D114480')]]

pair = pd.DataFrame()
for i in disease_id:
    for j in drug_id : 
        if didr[j].ix[i] == 1:
            temp_df = pd.DataFrame([[i, j]], columns = ['disease_id', 'drug_id'])
            pair = pair.append(temp_df)
            

pair.to_csv("/home/share/aschoi/data/Drug_Repositioning/7_new_association/bi_random_walk_Cdataset_association.tsv", sep='\t', index=False)

print len(pair)
pair.head(1)

common_feature = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/5_common_feature.tsv", header = None)
drugs_with_common_feature = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/6_drugs_with_common_feature.tsv", header =None)

common_feature = common_feature.rename(columns = {0:'umls_id'})
drugs_with_common_feature = drugs_with_common_feature.rename(columns = {0:'DrugBank_ID'})

filterd_drug = dr[dr['drug_id'].isin(drugs_with_common_feature['DrugBank_ID'])]
print len(filterd_drug)
filterd_drug.head(1)

filtered_asso = pair[pair['drug_id'].isin(filterd_drug['drug_id'])]
print len(filtered_asso)
filtered_asso.head()

len(filtered_asso.drop_duplicates(subset='disease_id'))

filtered_asso.to_csv("/home/share/aschoi/data/Drug_Repositioning/7_new_association/5_filtered_association(drug).tsv", sep="\t", index=False)
filtered_asso.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/7_new_association/5_filtered_association(drug).tsv", sep="\t", index=False)

df_sider = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/3_3_SIDER_STITCHid_DrugBankid.tsv")
df_offside = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/4_3_OFFSIDE_STITCHid_DrugBankid.tsv")

print(len(df_sider))
df_sider.head()

print(len(df_offside))
df_offside.head()

sider_drug = dr[dr['drug_id'].isin(df_sider['DrugBank_ID'])]
print len(sider_drug)
offside_drug = dr[dr['drug_id'].isin(df_offside['DrugBank_ID'])]
print len(offside_drug)

a = pd.concat([df_sider['DrugBank_ID'], df_offside['DrugBank_ID']]) # sider의 drug와 offside의 drug 합친거
b = a.drop_duplicates() # 중복제거

print len(df_sider['DrugBank_ID'].drop_duplicates())
print len(df_offside['DrugBank_ID'].drop_duplicates())
print len(a)
print len(b)

print len(dr)
print len(dr[dr['drug_id'].isin(b)]) # 제공되는 dataset중 feature가 존재하는 drug 수

print len(pair[pair['drug_id'].isin(df_sider['DrugBank_ID'])].drop_duplicates(subset='drug_id'))
pair[pair['drug_id'].isin(df_sider['DrugBank_ID'])].drop_duplicates(subset='drug_id').head()

print len(pair[pair['drug_id'].isin(df_offside['DrugBank_ID'])].drop_duplicates(subset='drug_id'))
pair[pair['drug_id'].isin(df_offside['DrugBank_ID'])].drop_duplicates(subset='drug_id').head()

print len(pair['drug_id'].drop_duplicates())
print len(pair[pair['drug_id'].isin(b)].drop_duplicates(subset='drug_id'))
print len(pair)
print len(pair[pair['drug_id'].isin(b)])
pair[pair['drug_id'].isin(b)].drop_duplicates(subset='drug_id').head()

c = pair[pair['drug_id'].isin(b)] # cdataset의 association중, feature가 존재하는 drug 찾은것.
print len(c)
print(len(pair.drop_duplicates(subset='disease_id')))
print(len(c.drop_duplicates(subset='disease_id')))

print len(b)

print len(c)
c.head(1)

len(c['drug_id'].drop_duplicates())

print len(pair['disease_id'].drop_duplicates())
print len(disease_id)

c.to_csv("/home/share/aschoi/data/Drug_Repositioning/7_new_association/5_before_gold_drug.tsv", sep='\t', index=False)

c.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/7_new_association/5_before_gold_drug.tsv", sep='\t', index=False)
