==> Drug_Repositioning/5_new_drug/5_drug(SIDER, OFFSIDE)_UMLS_sideeffect_names.ipynb <==
import csv
import numpy as np
import pandas as pd
import time
from urllib2 import urlopen # python2
import urllib2

df_sider_db = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/3_3_SIDER_STITCHid_DrugBankid.tsv")
df_offside_db = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/4_3_OFFSIDE_STITCHid_DrugBankid.tsv")

raw_sider = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/meddra_all_label_se_header.tsv")
raw_sider_pt = raw_sider[raw_sider.MedDRA_concept_type == 'PT']
sider = raw_sider_pt[['STITCH_compound_id_flat','STITCH_compound_id_stereo', 'UMLS_concept_id_for_MedDRA', 'side_effect_name']]

raw_offisde = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/3003377s-offsides.tsv")
offside = raw_offisde[['stitch_id','drug', 'umls_id', 'event']]

print len(sider)
sider.head(1)

df_sider_db.head(1)

offside.head(1)

df_offside_db.head(1)

raw_comb_umls = pd.concat([sider['UMLS_concept_id_for_MedDRA'], offside['umls_id']])
print len(sider['UMLS_concept_id_for_MedDRA'].drop_duplicates())
print len(offside['umls_id'].drop_duplicates())4
print len(raw_comb_umls.drop_duplicates())

raw_comb_name = pd.concat([sider['side_effect_name'], offside['event']])
print len(sider['side_effect_name'].drop_duplicates())
print len(offside['event'].drop_duplicates())
print len(raw_comb_name.drop_duplicates())

len(df_sider_db)

result = pd.DataFrame()
for index,row in df_sider_db.iterrows():
    record = row.tolist()
    found = sider[(sider.STITCH_compound_id_flat == record[0]) & (sider.STITCH_compound_id_stereo == record[1])]
    result = result.append(found)

result2 = pd.DataFrame()
for index2, row2 in df_offside_db.iterrows():
    record2 = row2.tolist()
    found2 = offside[offside.stitch_id == record2[0]]
    result2 = result2.append(found2)

result_fi = result.drop_duplicates()
result_fi.index = range(0, len(result_fi))
result_fi.to_csv("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/3_4_SIDER_db_side_effect.tsv", sep='\t', index=False)

print len(result2)
print len(result2.drop_duplicates())

result2_fi = result2.drop_duplicates()
result2_fi.index = range(0, len(result2_fi))
result2_fi.to_csv("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/4_4_OFFSIDE_db_side_effect.tsv", sep='\t', index=False)

result_fi.head()
