==> Drug_Repositioning/5_new_drug/6_SIDER_OFFSIDE_combine_by_umlsID.ipynb <==
import csv
import numpy as np
import os
import pandas as pd
import time

# SIDER랑 OFFSIDE에 등록된 compound의 DrugBank id 정보
df_sider_id = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/3_3_SIDER_STITCHid_DrugBankid.tsv")
df_offside_id = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/4_3_OFFSIDE_STITCHid_DrugBankid.tsv")
# SIDER랑 OFFSIDE에 등록된 compound의 side effects 정보 (umls id로 표현되어 있음.)
df_sider_umls = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/3_4_SIDER_db_side_effect.tsv")
df_offside_umls = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/4_4_OFFSIDE_db_side_effect.tsv")

print len(df_sider_id)
df_sider_id.head(1)

print len(df_sider_umls)
df_sider_umls.head(1)

df_sider = pd.merge(left = df_sider_id, right = df_sider_umls)
df_sider.head(1)

print len(df_sider)
print len(df_sider.drop_duplicates())
print len(df_sider.drop_duplicates(subset = 'DrugBank_ID'))
print len(df_sider.drop_duplicates(subset = ['STITCH_compound_id_flat', 'STITCH_compound_id_stereo', 'DrugBank_ID']))
print len(df_sider.drop_duplicates(subset = ['STITCH_compound_id_flat', 'STITCH_compound_id_stereo', 'UMLS_concept_id_for_MedDRA', 'side_effect_name']))

print len(df_offside_id)
print len(df_offside_id.drop_duplicates(subset = 'DrugBank_ID'))
df_offside_id.head(1)

print len(df_offside_umls)
df_offside_umls.head(1)

df_offside = pd.merge(left = df_offside_id, right = df_offside_umls, left_on = 'STITCH_compound_id_stereo', right_on = 'stitch_id')
df_offside.head(1)

print len(df_offside)
print len(df_offside.drop_duplicates(subset = 'DrugBank_ID'))
print len(df_offside.drop_duplicates(subset = ['STITCH_compound_id_stereo', 'DrugBank_ID']))
print len(df_offside.drop_duplicates(subset = ['stitch_id', 'drug', 'umls_id', 'event']))

all_umls = pd.concat([df_sider_umls['UMLS_concept_id_for_MedDRA'], df_offside_umls['umls_id']])
unique_umls = all_umls.drop_duplicates()

print len(df_sider_umls['UMLS_concept_id_for_MedDRA'].drop_duplicates())
print len(df_offside_umls['umls_id'].drop_duplicates())
print len(all_umls)
print len(unique_umls)

umls_inter = pd.Series(list(set(df_sider_umls['UMLS_concept_id_for_MedDRA']).intersection(set(df_offside_umls['umls_id']))))
len(umls_inter)

# SIDER에서 공통된 feature를 갖고 있는 compound
dup_free_sider = df_sider[df_sider['UMLS_concept_id_for_MedDRA'].isin(umls_inter)].drop_duplicates(subset = 'DrugBank_ID')
print len(dup_free_sider)
dup_free_sider.head(1)

# OFFSIDE에서 공통된 feature를 갖고 있는 compound
dup_free_off = df_offside[df_offside['umls_id'].isin(umls_inter)].drop_duplicates(subset = 'DrugBank_ID')
print len(dup_free_off)
dup_free_off.head(1)

#SIDER와 OFFSIDE의 공통 feature를 가지는 drug들
drug_with_common_feature = pd.concat([dup_free_sider['DrugBank_ID'], dup_free_off['DrugBank_ID']]).drop_duplicates()

print len(drug_with_common_feature)
drug_with_common_feature.head()

umls_inter.to_csv("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/5_common_feature.tsv", sep='\t', index=False)
umls_inter.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/5_common_feature.tsv", sep='\t', index=False)
drug_with_common_feature.to_csv("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/6_drugs_with_common_feature.tsv", sep='\t', index=False)
drug_with_common_feature.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/6_drugs_with_common_feature.tsv", sep='\t', index=False)

# gold drug & features

gold = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/7_new_association/7_gold_association.tsv")

gold_drug_id = gold['drug_id'].drop_duplicates()
print len(gold_drug_id)

# SIDER에서 공통된 feature를 갖고 있는 drug들 (feautre정보 포함)
sider_feature = df_sider[df_sider['UMLS_concept_id_for_MedDRA'].isin(umls_inter)]
print len(sider_feature)
print len(sider_feature.drop_duplicates(subset = "DrugBank_ID"))

# 이 중, gold standard drug에 해당하는 것.
sider_gold = sider_feature[sider_feature['DrugBank_ID'].isin(gold_drug_id)][['DrugBank_ID', 'UMLS_concept_id_for_MedDRA', 'side_effect_name']]
print len(sider_gold)
sider_gold.head(1)
print len(sider_gold.drop_duplicates(subset = 'DrugBank_ID'))

# OFFSIDE에서 공통된 feature를 갖고 있는 drug들 (feature 정보 포함)
offside_feature = df_offside[df_offside['umls_id'].isin(umls_inter)]
print len(offside_feature)
print len(offside_feature.drop_duplicates(subset = 'DrugBank_ID'))

# 이 중, gold standard drug에 해당하는 것.
offside_gold = offside_feature[offside_feature['DrugBank_ID'].isin(gold_drug_id)][['DrugBank_ID', 'umls_id', 'event']]
print len(offside_gold)
offside_gold.head(1)
print len(offside_gold.drop_duplicates(subset = "DrugBank_ID"))

sider_gold = sider_gold.rename(columns = {'UMLS_concept_id_for_MedDRA' : 'umls_id', 'side_effect_name' : 'event'})
sider_gold.head(1)

# sider의 offside의 gold drug, side effects의 umls_id, 이름.
frames = [sider_gold, offside_gold]
combined = pd.concat(frames)

print len(combined)
print len(combined.drop_duplicates())
print len(combined.drop_duplicates(subset = ['DrugBank_ID', 'umls_id']))
print len(combined.drop_duplicates(subset = "DrugBank_ID"))
print len(combined.drop_duplicates(subset= "umls_id"))
combined[combined.duplicated(subset = ['DrugBank_ID', 'umls_id'],keep = False)].sort_values(by = ['DrugBank_ID','umls_id']).head()

combined_name = combined.drop_duplicates(subset = ['DrugBank_ID', 'umls_id'])
print len(combined_name)
combined_name.to_csv("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/7_gold_drug_feature_name.tsv", sep="\t", index=False)
combined_name.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/7_gold_drug_feature_name.tsv", sep="\t", index=False)

frames2 = [sider_gold[['DrugBank_ID', 'umls_id']], offside_gold[['DrugBank_ID', 'umls_id']]]
combined2 = pd.concat(frames2)

print len(combined2)
combined_id = combined2.drop_duplicates()
print len(combined_id)
combined_id.to_csv("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/7_2_gold_drug_feature_id.tsv", sep="\t", index=False)
combined_id.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/7_2_gold_drug_feature_id.tsv", sep="\t", index=False)


#여기서 association에 포함되는 drug의 수는?

befor_gold = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/7_new_association/5_before_gold_drug.tsv")
asso = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/7_new_association/4_bi_random_walk_Cdataset_association.tsv")

print len(asso)
asso.head(1)

filtered_asso = asso[asso['drug_id'].isin(drug_with_common_feature)]

print len(filtered_asso)
filtered_asso.head()

print len(filtered_asso.drop_duplicates(subset='disease_id'))
print len(filtered_asso.drop_duplicates(subset='drug_id'))

print len(dup_free_sider[dup_free_sider['DrugBank_ID'].isin(befor_gold['drug_id'])])
dup_free_sider[dup_free_sider['DrugBank_ID'].isin(befor_gold['drug_id'])].head(1)

# 질병과의 association이 존재하rh, drug feature를 가진 drug가 sider
print len(dup_free_off[dup_free_off['DrugBank_ID'].isin(befor_gold['drug_id'])])
dup_free_off[dup_free_off['DrugBank_ID'].isin(befor_gold['drug_id'])].head(1)

si = dup_free_sider[dup_free_sider['DrugBank_ID'].isin(befor_gold['drug_id'])]['DrugBank_ID']

off = dup_free_off[dup_free_off['DrugBank_ID'].isin(befor_gold['drug_id'])]['DrugBank_ID']

len(pd.Series(list(set(si).intersection(set(off)))))
