==> Drug_Repositioning/5_new_drug/7_herb_compound_parsing.ipynb <==
import MySQLdb
import time
import pandas as pd
import numpy as np
import time

#접속하기
db_handle = MySQLdb.connect('localhost', 'root','comsysbio', 'COCONUTv5')
cur = db_handle.cursor(MySQLdb.cursors.DictCursor) # query를 통해 데이터를 얻을 때 python의 dict로 얻기 위한 방법
cursor = db_handle.cursor()

compound = pd.read_table("/home/share/aschoi/nas/users/asolchoi/data/COCONUT_Herbal/6_Herb/COCONUT_herb_compound(remove duplicate)+ref2_exclude_KTKP_include(TCMID,TCM-ID,KAMPO).txt")

print len(compound)
compound.head(1)

########################Herb compound Phenotype#####################
#tuple type
df_compound_pheno = pd.DataFrame()
for i in compound['compID']:
    compID = i
    query2 = "SELECT * from Compound_Phenotype where compID = " + compID
    #query1 = "SELECT compID, ph From Compound WHERE compID IN(SELECT compID FROM Herb_Compound WHERE herbID= %ld)" % herb_id
    cursor.execute(query2)
    results = cursor.fetchall()
    if results != ():
        for j in results:
            temp_df = pd.DataFrame([[j[1], j[2], j[3], j[5], j[7]]], columns = ['compID', 'phenID', 'effectType', 'evidence', 'reference'])
            df_compound_pheno = df_compound_pheno.append(temp_df)

treatment = df_compound_pheno[df_compound_pheno.effectType == 'Treatment']
print len(treatment)

treatment.to_csv("/home/share/aschoi/data/COCONUT_Herbal/6_Herb/5_1_herb_compound_phenotype_treatment.tsv", sep="\t", index=False)
treatment.to_csv("/home/share/aschoi/nas/users/asolchoi/data/COCONUT_Herbal/6_Herb/5_1_herb_compound_phenotype_treatment.tsv", sep="\t", index=False)

association = df_compound_pheno[df_compound_pheno.effectType == 'Association']
print len(association)

association.to_csv("/home/share/aschoi/data/COCONUT_Herbal/6_Herb/5_1_herb_compound_phenotype_association.tsv", sep="\t", index=False)
association.to_csv("/home/share/aschoi/nas/users/asolchoi/data/COCONUT_Herbal/6_Herb/5_1_herb_compound_phenotype_association.tsv", sep="\t", index=False)

cause = df_compound_pheno[df_compound_pheno.effectType == 'Cause']
print len(cause)

cause.to_csv("/home/share/aschoi/data/COCONUT_Herbal/6_Herb/5_1_herb_compound_phenotype_cause.tsv", sep="\t", index=False)
cause.to_csv("/home/share/aschoi/nas/users/asolchoi/data/COCONUT_Herbal/6_Herb/5_1_herb_compound_phenotype_cause.tsv", sep="\t", index=False)

print len(df_compound_pheno)
print len(df_compound_pheno.drop_duplicates(subset = 'phenID'))
print len(df_compound_pheno.drop_duplicates(subset = ['compID', 'phenID']))
df_compound_pheno.head()

########################Herb compound Phenotype umls id#####################
#tuple type
df_compound_pheno_umls = pd.DataFrame()
for index, row in df_compound_pheno.iterrows():
    phenID = row['phenID']
    query2 = "SELECT umlsID from Phenotype where phenID = " + str(phenID)
    cursor.execute(query2)
    results = cursor.fetchall()
    if results != ():
        row['umlsID'] = results[0][0]
        df_compound_pheno_umls = df_compound_pheno_umls.append(row)
    else:
        print("empty")
        break

print len(df_compound_pheno_umls)
df_compound_pheno_umls.head()

treatment2 = df_compound_pheno_umls[df_compound_pheno_umls.effectType == 'Treatment']
print len(treatment2)

treatment2.to_csv("/home/share/aschoi/data/COCONUT_Herbal/6_Herb/5_2_herb_compound_phenotype_treatment_umls.tsv", sep="\t", index=False)
treatment2.to_csv("/home/share/aschoi/nas/users/asolchoi/data/COCONUT_Herbal/6_Herb/5_2_herb_compound_phenotype_treatment_umls.tsv", sep="\t", index=False)

association2 = df_compound_pheno_umls[df_compound_pheno_umls.effectType == 'Association']
print len(association2)

association2.to_csv("/home/share/aschoi/data/COCONUT_Herbal/6_Herb/5_1_herb_compound_phenotype_association_umls.tsv", sep="\t", index=False)
association2.to_csv("/home/share/aschoi/nas/users/asolchoi/data/COCONUT_Herbal/6_Herb/5_1_herb_compound_phenotype_association_umls.tsv", sep="\t", index=False)

cause2 = df_compound_pheno_umls[df_compound_pheno_umls.effectType == 'Cause']
print len(cause)
cause2.to_csv("/home/share/aschoi/data/COCONUT_Herbal/6_Herb/5_2_herb_compound_phenotype_cause_umls.tsv", sep="\t", index=False)
cause2.to_csv("/home/share/aschoi/nas/users/asolchoi/data/COCONUT_Herbal/6_Herb/5_2_herb_compound_phenotype_cause_umls.tsv", sep="\t", index=False)
