==> Drug_Repositioning/5_new_drug/8_herb_compound_cause_association.ipynb <==
import csv
import numpy as np
import os
import pandas as pd
import time

cause = pd.read_table("/home/share/aschoi/data/COCONUT_Herbal/6_Herb/5_2_herb_compound_phenotype_cause_umls.tsv")

print len(cause)
cause.head(1)

len(cause.drop_duplicates(subset ='compID'))

common_feature = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/5_new_drug/5_common_feature.tsv", header=None)
common_feature = common_feature.rename(columns = {0 : 'umlsID'})

len(cause.drop_duplicates(subset = 'umlsID'))

common_feature.head()

print len(common_feature[common_feature['umlsID'].isin(cause['umlsID'])])
print len(cause[cause['umlsID'].isin(common_feature['umlsID'])])
print len(cause[cause['umlsID'].isin(common_feature['umlsID'])].drop_duplicates(subset='umlsID'))
print len(cause[cause['umlsID'].isin(common_feature['umlsID'])].drop_duplicates(subset = 'compID'))
cause[cause['umlsID'].isin(common_feature['umlsID'])].drop_duplicates(subset = 'compID').head()

asso = pd.read_table("/home/share/aschoi/data/COCONUT_Herbal/6_Herb/5_2_herb_compound_phenotype_association_umls.tsv")

print len(asso)
asso.head()

print len(asso)
print len(asso.drop_duplicates(subset = 'compID'))
print len(asso.drop_duplicates(subset = 'umlsID'))
print len(asso.drop_duplicates(subset = ['umlsID', 'compID']))

disease = pd.read_table("/home/share/aschoi/data/Drug_Repositioning/7_new_association/4_bi_random_walk_Cdataset_association.tsv")

print len(disease)
disease.head()

mapping = pd.read_excel("/home/share/aschoi/data/Drug_Repositioning/7_new_association/PREDICT_S6 Mapping between OMIM disease and UMLS concepts.xls")

print len(mapping)
mapping.head()

di_mimNum = disease['disease_id'].drop_duplicates().str[1:]
int_di_mimNum = list()
for i in di_mimNum:
    int_di_mimNum.append (int(i))

mapped = mapping[mapping['OMIM ID'].isin(int_di_mimNum)]
print len(mapped)
mapped.head()

print len(mapped[mapped['UMLS concept ID'].isin(asso['umlsID'])])
print len(mapped[mapped['UMLS concept ID'].isin(asso['umlsID'])].drop_duplicates(subset = 'UMLS concept ID'))
print len(asso[asso['umlsID'].isin(mapped['UMLS concept ID'])].drop_duplicates(subset = 'umlsID'))
print len(asso[asso['umlsID'].isin(mapped['UMLS concept ID'])])
print len(asso[asso['umlsID'].isin(mapped['UMLS concept ID'])].drop_duplicates(subset = 'compID'))

asso[asso['umlsID'].isin(mapped['UMLS concept ID'])]

cause_compID = cause[cause['umlsID'].isin(common_feature['umlsID'])].drop_duplicates(subset = 'compID')['compID']

asso_compID = asso[asso['umlsID'].isin(mapped['UMLS concept ID'])].drop_duplicates(subset = 'compID')['compID']

len(asso_compID[asso_compID.isin(cause_compID)])

cause_compID = cause[cause['umlsID'].isin(common_feature['umlsID'])]['compID']
asso_compID = asso[asso['umlsID'].isin(mapped['UMLS concept ID'])]
asso_compID[asso_compID['compID'].isin(cause_compID)].drop_duplicates(subset='umlsID')
