==> Drug_Repositioning/5_new_drug/9_herb_compound_SIDER_OFFSIDE_application_dataset.ipynb <==
import csv
import numpy as np
import os
import pandas as pd
import requests
import time

####### herb compound의 stitch ID를 mapping하는 작업. pubchem ID를 매개로 사용할 것이다.

sider_cid = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/3_0_stitch(sider)_pubchem.tsv")
herb_cid = pd.read_table("/DAS_Storage1/aschoi/data/COCONUT_Herbal/7_Ezyme_tool/2-5_herb_pubchem.tsv")

print len(sider_cid)
sider_cid.head(1)

print len(herb_cid)
herb_cid.head(1)

a = sider_cid[sider_cid['pubchem_cid'].isin(herb_cid['pubchemID'])]
print len(a)
print len(a.drop_duplicates(subset = ['stitch_flat', 'stitch_stereo']))
a.head(2)

b = herb_cid[herb_cid['pubchemID'].isin(sider_cid['pubchem_cid'])]
print len(b)
print len(b.drop_duplicates(subset = 'compID'))
b.head(2)

result = pd.DataFrame()
for index, row in b.iterrows():
    pubchem_id = row['pubchemID']
    filtered = sider_cid[sider_cid.pubchem_cid == pubchem_id]
    if filtered.empty:
        continue
    else:
        for index2, row2 in filtered.iterrows():
            flat = row2['stitch_flat']
            stereo = row2['stitch_stereo']
            temp_df = pd.DataFrame([[row['compID'], row['compName'], flat, stereo]], 
                                  columns = ['compID', 'compName', 'stitch_flat', 'stitch_stereo'])
            result = result.append(temp_df)

print len(result)
print len(result.drop_duplicates())

result.head()

result.drop_duplicates().to_csv("/DAS_Storage1/aschoi/data/COCONUT_Herbal/6_Herb/6_1_herb_stitchID(SIDER).tsv", sep='\t', index=False)
result.drop_duplicates().to_csv("/home/share/aschoi/nas/users/asolchoi/data/COCONUT_Herbal/6_Herb/6_1_herb_stitchID(SIDER).tsv", sep='\t', index=False)

######## herb compound의 바꾼 stitch id를 사용. SIDER에서 해당하는 side effect를 뽑아낸다.

raw_sider = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/meddra_all_label_se_header.tsv")

raw_sider_pt = raw_sider[raw_sider.MedDRA_concept_type == 'PT']
print(len(raw_sider_pt))
raw_sider_pt.head(1)

sider = raw_sider_pt[['STITCH_compound_id_flat','STITCH_compound_id_stereo', 'UMLS_concept_id_for_MedDRA', 'side_effect_name']]

sider_wo_dup = sider.drop_duplicates()
print(len(sider_wo_dup))
sider_wo_dup.head()

sider_result = pd.DataFrame()
for index, row in result.iterrows():
    flat = row['stitch_flat']
    stereo = row['stitch_stereo']
    filtered = sider_wo_dup[(sider_wo_dup.STITCH_compound_id_flat == flat) & (sider_wo_dup.STITCH_compound_id_stereo == stereo)]
    if filtered.empty:
        print flat, stereo
        continue
    else:
        for index2, row2 in filtered.iterrows():
            umls_id = row2['UMLS_concept_id_for_MedDRA']
            side_effect = row2['side_effect_name']
            temp_df = pd.DataFrame([[row['compID'], row['compName'], umls_id, side_effect]], 
                                   columns = ['compID', 'compName', 'umls_id', 'side_effect_name'])
            sider_result = sider_result.append(temp_df)

print len(sider_result)
print len(sider_result.drop_duplicates())

sider[sider.duplicated(keep=False)].sort_values(by = ['STITCH_compound_id_flat', 'STITCH_compound_id_stereo','UMLS_concept_id_for_MedDRA']).head()

sider_result = sider_result.drop_duplicates()

print len(sider_result.drop_duplicates(subset = 'compID')) # SIDER의 side effect를 가지고 있는 herb compound의 수
print len(sider_result.drop_duplicates(subset = 'umls_id')) # SIDER의 side effect의 전체 수
sider_result.head()

sider_result.to_csv("/DAS_Storage1/aschoi/data/COCONUT_Herbal/6_Herb/6_2_herb_sider_sideEffect.tsv", sep='\t', index=False)
sider_result[['compID', 'compName', 'umls_id']].to_csv("/DAS_Storage1/aschoi/data/COCONUT_Herbal/6_Herb/6_3_herb_sider_umlsID.tsv", sep='\t', index=False)
sider_result.to_csv("/home/share/aschoi/nas/users/asolchoi/data/COCONUT_Herbal/6_Herb/6_2_herb_sider_sideEffect.tsv", sep='\t', index=False)
sider_result[['compID', 'compName', 'umls_id']].to_csv("/home/share/aschoi/nas/users/asolchoi/data/COCONUT_Herbal/6_Herb/6_3_herb_sider_umlsID.tsv", sep='\t', index=False)

########## 여기서부터는 OFFSIDE

stitch_cid = pd.read_table("/DAS_Storage1/aschoi/data/COCONUT_Herbal/7_Ezyme_tool/STITCH_CID.tsv")

print len(stitch_cid)
stitch_cid.head(2)

raw_offisde = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/3003377s-offsides.tsv")

print(len(raw_offisde))
offside = raw_offisde[['stitch_id','drug', 'umls_id', 'event']]
offside.head(2)

g = stitch_cid[stitch_cid.stereo_chemical.isin(raw_offisde['stitch_id'])]
print len(g)
print len(g.drop_duplicates(subset = 'flat_chemical')) # 여기서 왜 flat chemical을 하는지 모름...
g.head(1)

# OFFSIDE의 stitch id과 pubchem id를 매칭해놓은것.
g.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/4_0_stitch(offside)_pubchem.tsv")
g.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/4_0_stitch(offside)_pubchem.tsv")

print len(g[g.source_id.isin(herb_cid['pubchemID'])]) 
print len(g[g.source_id.isin(herb_cid['pubchemID'])].drop_duplicates(subset = 'stereo_chemical')) # offside 정보를 가지고 있는 herb compound
g[g.source_id.isin(herb_cid['pubchemID'])].head(2)

h = herb_cid[herb_cid['pubchemID'].isin(g['source_id'])] 
print len(h)
print len(h.drop_duplicates(subset = 'compID'))# herb compound 중, offisde 정보를 가지고 있는 herb compound의 수 (위아래가 왜 수가 다른지는 모를..)
h.head(2)

# herb copmound와 offside id를 매핑하는 과정 (pubchem id 사용)
result2 = pd.DataFrame()
for index, row in h.iterrows():
    pubchem_id = row['pubchemID']
    filtered = g[g.source_id == pubchem_id]
    if filtered.empty:
        continue
    else:
        for index2, row2 in filtered.iterrows():
            stereo = row2['stereo_chemical']
            temp_df = pd.DataFrame([[row['compID'], row['compName'], stereo]], 
                                  columns = ['compID', 'compName','stitch_stereo'])
            result2 = result2.append(temp_df)

# 위에서 바꾼 offside id를 통해 herb compound마다 adr을 찾는다. 
offside_result = pd.DataFrame()
for index, row in result2.iterrows():
    stereo = row['stitch_stereo']
    filtered = raw_offisde[raw_offisde.stitch_id == stereo]
    if filtered.empty:
        print stereo
        continue
    else:
        for index2, row2 in filtered.iterrows():
            umls_id = row2['umls_id']
            side_effect = row2['event']
            temp_df = pd.DataFrame([[row['compID'], row['compName'], umls_id, side_effect]], 
                                   columns = ['compID', 'compName', 'umls_id', 'side_effect_name'])
            offside_result = offside_result.append(temp_df)

print len(offside_result)
print len(offside_result.drop_duplicates())
print len(offside_result.drop_duplicates(subset = 'compID')) # offside의 정보를 갖고 있는 herb compound의 수
print len(offside_result.drop_duplicates(subset ='umls_id')) # 그것의 adr의 수
offside_result.head(2)

offside_result.to_csv("/DAS_Storage1/aschoi/data/COCONUT_Herbal/6_Herb/6_2_herb_offside_sideEffect.tsv", sep = '\t', index = False)
offside_result[['compID', 'compName', 'umls_id']].to_csv("/DAS_Storage1/aschoi/data/COCONUT_Herbal/6_Herb/6_3_herb_offside_umlsID.tsv", sep = '\t', index = False)
offside_result.to_csv("/home/share/aschoi/nas/users/asolchoi/data/COCONUT_Herbal/6_Herb/6_2_herb_offside_sideEffect.tsv", sep = '\t', index = False)
offside_result[['compID', 'compName', 'umls_id']].to_csv("/home/share/aschoi/nas/users/asolchoi/data/COCONUT_Herbal/6_Herb/6_3_herb_offside_umlsID.tsv", sep = '\t', index = False)

# herb_compound가 갖고 있는 side effect feature와 gold drug의 common feature 비교해서 걸러내기

drug = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/8_gold_standard_drug_binary_matrix.tsv", index_col = 0)

common_feature = pd.Series(drug.index.values)
print len(common_feature)

len(sider_result[sider_result['umls_id'].isin(common_feature)].drop_duplicates(subset = 'umls_id'))

len(offside_result[offside_result['umls_id'].isin(common_feature)].drop_duplicates(subset = 'umls_id'))

# 기존 COCONUT database에서 제공하는 정보도 합쳐야함.

herb_cause = pd.read_table("/DAS_Storage1/aschoi/data/COCONUT_Herbal/6_Herb/5_2_herb_compound_phenotype_cause_umls.tsv")

len(herb_cause[herb_cause['compID'].isin(sider_result['compID'])].drop_duplicates(subset = 'compID'))

coconut = herb_cause[['compID', 'umlsID']]
coconut = coconut.rename(columns = {'umlsID' : 'umls_id'})
sider = sider_result[['compID', 'umls_id']] 
offside = offside_result[['compID', 'umls_id']]
frames = [coconut, sider, offside]
combined = pd.concat(frames)

combined.head(1)

print len(combined.drop_duplicates(subset = 'compID')) # herb compound 의 수
print len(combined.drop_duplicates(subset = 'umls_id')) # 그 herb compound가 가지고 있는 side effect수

print len(combined[combined['umls_id'].isin(common_feature)].drop_duplicates(subset = 'compID')) # gold standard drug feature를 가지고 있는 herb compound의 수
print len(combined[combined['umls_id'].isin(common_feature)].drop_duplicates(subset = 'umls_id')) # gold standard drug feature와 겹치는 feature의 수
k = combined[combined['umls_id'].isin(common_feature)]

k.head(2)

k.to_csv("/DAS_Storage1/aschoi/data/COCONUT_Herbal/6_Herb/7_herb_compound_combined_common_phenotypic_effects.tsv", sep='\t', index=False)
k.to_csv("/home/share/aschoi/nas/users/asolchoi/data/COCONUT_Herbal/6_Herb/7_herb_compound_combined_common_phenotypic_effects.tsv", sep='\t', index=False)

### learning이 가능한 herb compound(바로 위에서 구한 k)와 disease의 association

herb_asso = pd.read_table("/DAS_Storage1/aschoi/data/COCONUT_Herbal/6_Herb/5_2_herb_compound_phenotype_association_umls.tsv")

herb_asso.head(1)

print len(herb_asso[herb_asso['compID'].isin(k['compID'])].drop_duplicates())  # row의 수가 learning이 가능한 herb compound와 disease의 association 수
herb_asso[herb_asso['compID'].isin(k['compID'])].drop_duplicates().head()

learning_herb_asso = herb_asso[herb_asso['compID'].isin(k['compID'])].drop_duplicates()
learning_herb_asso.to_csv("/DAS_Storage1/aschoi/data/COCONUT_Herbal/6_Herb/7_2_learning_herb_compound_disease_association.tsv", sep='\t', index=False)
learning_herb_asso.to_csv("/home/share/aschoi/nas/users/asolchoi/data/COCONUT_Herbal/6_Herb/7_2_learning_herb_compound_disease_association.tsv", sep='\t', index=False)

# 이건 coconut에 있는 feature로만 찾은 herb compound -disease association임
print len(herb_asso[herb_asso['compID'].isin(herb_cause['compID'])].drop_duplicates())
herb_asso[herb_asso['compID'].isin(herb_cause['compID'])].drop_duplicates()

추가 작업 필요함 -> association 이 있다고 하는 disease의 feature를 구하고(DisNet에서)
이 feature와 gold standard disease feature와 공통을 구해서 사용가능한 association을 filter해야함.
