==> Drug_Repositioning/5_new_drug/10_similarity_scores(PREDICT)-Copy1.ipynb <==
import csv
import numpy as np
import os
import pandas as pd
import time

gold_drug = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/7_2_gold_drug_feature_id.tsv")

gold_drug_id = gold_drug['DrugBank_ID'].drop_duplicates()

drugbank = pd.read_table("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/drug_set.txt", header=None)
drugbank = drugbank.rename(columns = {0:'DrugBank_ID', 1:'Drug_Name', 2:'InChI', 3:'SMILES', 4:'State'})
drugbank.head(1)

df_gold_drug_smiles = drugbank[drugbank['DrugBank_ID'].isin(gold_drug_id)]
sorter = gold_drug_id.tolist()
df_gold_drug_smiles.DrugBank_ID = df_gold_drug_smiles.DrugBank_ID.astype("category")
df_gold_drug_smiles.DrugBank_ID.cat.set_categories(sorter, inplace=True)
df_gold_drug_smiles = df_gold_drug_smiles.sort_values(["DrugBank_ID"])
print len(df_gold_drug_smiles)
df_gold_drug_smiles.head(2) # drug bank id로 중복체크 했을 때 중복 없었음.

import pybel
df_sim = pd.DataFrame()
drug_id = df_gold_drug_smiles['DrugBank_ID'].tolist()
for i in drug_id:
    smi1 = df_gold_drug_smiles[df_gold_drug_smiles.DrugBank_ID == i]['SMILES'].tolist()[0]
    for j in drug_id:
        smi2 = df_gold_drug_smiles[df_gold_drug_smiles.DrugBank_ID == j]['SMILES'].tolist()[0]
        
        smiles = [smi1, smi2]
        mol = [pybel.readstring('smi',x) for x in smiles]
        fps = [x.calcfp(fptype='FP2') for x in mol]
        sim = fps[0] | fps[1]
        
        temp_df = pd.DataFrame([[i, j, sim]], columns = ['DrugBank_ID1','DrugBank_ID2', 'Similarity_Score'])
        df_sim = df_sim.append(temp_df)

df_sim.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/9_gold_drug_similarity_structure.tsv", sep='\t', index=False)
df_sim.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/9_gold_drug_similarity_structure.tsv", sep='\t', index=False)

df_sim = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/9_gold_drug_similarity_structure.tsv")

gold_drug_matrix = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/8_gold_standard_drug_binary_matrix.tsv", index_col=0)

gold_drug_matrix.head()

from sklearn.metrics import jaccard_similarity_score
df_jac = pd.DataFrame()
drug_id = df_gold_drug_smiles['DrugBank_ID'].tolist()
for i in drug_id:
    desc1 = gold_drug_matrix[i]
    for j in drug_id:
        desc2 = gold_drug_matrix[j]
        
        jac_sim = jaccard_similarity_score(desc1,desc2)
        
        temp_df = pd.DataFrame([[i, j, jac_sim]], columns = ['DrugBank_ID1','DrugBank_ID2', 'Similarity_Score'])
        df_jac = df_jac.append(temp_df)

df_jac.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/9_gold_drug_similarity_phenotype.tsv", sep='\t', index=False)
df_jac.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/9_gold_drug_similarity_phenotype.tsv", sep='\t', index=False)

df_jac = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/9_gold_drug_similarity_phenotype.tsv")

############ sequence based similarity
drug_target = pd.read_table("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/human2.txt", header=None)
target_sequence = pd.read_table("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/protein_seq.txt")

drug_target = drug_target.rename(columns={0:'DrugBank_ID', 1:'Target'})
print len(drug_target)
drug_target.head(2)

print len(target_sequence)
target_sequence.head(2)

drug_seq = pd.merge(left=drug_target, right=target_sequence, left_on='Target', right_on='Entry')
print len(drug_seq)
drug_seq.head(2)

import swalign
# choose your own values here… 2 and -1 are common.
match = 2
mismatch = -1
scoring = swalign.NucleotideScoringMatrix(match, mismatch)

sw = swalign.LocalAlignment(scoring)  # you can also choose gap penalties, etc...
alignment = sw.align('AAAAAA','AAAAAA')
alignment.dump()

len(queries)

import swalign
import math
from itertools import product

match = 2
mismatch = -1
scoring = swalign.NucleotideScoringMatrix(match, mismatch)
sw = swalign.LocalAlignment(scoring)  # you can also choose gap penalties, etc...

### sequence pair만들기
drug_id = df_gold_drug_smiles['DrugBank_ID'].tolist()
gold_drug_seq = drug_seq[drug_seq['DrugBank_ID'].isin(drug_id)]
queries = list()
print 'start : ' + time.strftime('%a %H:%M:%S')
for i in drug_id:
    seqs1 = gold_drug_seq[gold_drug_seq.DrugBank_ID == i]['Sequence'].tolist()
    for j in drug_id:
        if i==j:
            continue
        seqs2 = gold_drug_seq[gold_drug_seq.DrugBank_ID == j]['Sequence'].tolist()
        queries.append(list(product(seqs1, seqs2)))
print 'end : ' + time.strftime('%a %H:%M:%S')

###sequne pair의 sw score 구하기
df_sw = pd.DataFrame()
cnt = 0
scores = list()
print len(queries)
print 'start : ' + time.strftime('%a %H:%M:%S')
for k in queries:
    for j in k:
        seq1 = j[0]
        seq2 = j[1]
        sw_score = sw.align(seq1,seq2).score / (math.sqrt(sw.align(seq1, seq1).score)* math.sqrt(sw.align(seq2, seq2).score))
        scores.append(sw_score)
    final_score = max(scores) / len(scores)
    temp_df = pd.DataFrame([[i, j, final_score]], columns = ['DrugBank_ID1','DrugBank_ID2', 'Similarity_Score'])
    df_sw = df_sw.append(temp_df)

    cnt = cnt + 1
    if cnt %  1000 == 0:
        print 'cnt : {} '.format(cnt) + time.strftime('%a %H:%M:%S')
print time.strftime('%a %H:%M:%S')


cnt  # 중단한 시간은 수요일 오후 3시 41분


#문제가 많음.
1. 중복문제. a-b , b-a 가 제대로 삭제 되지 않은채로 평균값을 구하고 있음.
2. complexity가 너무 높음.
