==> Drug_Repositioning/5_new_drug/11_indep_similarities(PREDICT).ipynb <==
import csv
import numpy as np
import os
import pandas as pd
import time
import pybel

'''2017.4.20. gold에 포함되지 않는 indep drug와 gold drug간의 similarity 정보도 필요함.'''
indep_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/new_predict/3_indep_learning_asso.tsv")
print len(indep_asso['drug_id'].drop_duplicates())
gold_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/new_predict/1_gold_learning_asso.tsv")
print len(gold_asso['drug_id'].drop_duplicates())

gold_dr_id = gold_asso['drug_id'].drop_duplicates()
indep_dr_id = indep_asso['drug_id'].drop_duplicates()

drugbank = pd.read_table("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/drug_set.txt", header=None)
drugbank = drugbank.rename(columns = {0:'DrugBank_ID', 1:'Drug_Name', 2:'InChI', 3:'SMILES', 4:'State'})
drugbank.set_index(['DrugBank_ID'], inplace = True)
drugbank.head(1)

indep_sim_mat = pd.DataFrame(columns = gold_dr_id.tolist(), index = indep_dr_id.tolist())
indep_sim_mat = indep_sim_mat.fillna(0.0)
indep_sim_mat.head(1)

for index, row in indep_sim_mat.iterrows():
    indep_smi = drugbank.filter(items = [index], axis = 0)['SMILES'].tolist()[0]
    for dr_id in gold_dr_id:
        gold_smi = drugbank.filter(items = [dr_id], axis = 0)['SMILES'].tolist()[0]
        
        smiles = [indep_smi, gold_smi]
        mol = [pybel.readstring('smi',x) for x in smiles]
        fps = [x.calcfp(fptype='FP2') for x in mol]
        sim = fps[0] | fps[1]
        indep_sim_mat.set_value(index, dr_id, sim)  

indep_sim_mat.isnull().any().any()

indep_sim_mat.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/new_predict/2_2_2_indep_drug_gold_drug_str_sim_mat.tsv", sep='\t')
indep_sim_mat.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/new_predict/2_2_2_indep_drug_gold_drug_str_sim_mat.tsv", sep='\t')

'''2017.4.20. 끝'''

# indep의 structure sim descriptor 
indep_sim = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/new_predict/2_indep_drug_drug_str_sim_mat.tsv", index_col = 0)

print len(indep_sim)
indep_sim.head()

indep_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/9_2_indep_learning_association.tsv")
print len(indep_asso)
print len(indep_asso[indep_asso.association == 1]) # positive
print len(indep_asso[indep_asso.association == 0]) # negative
print len(indep_asso['drug'].drop_duplicates())
print len(indep_asso['disease'].drop_duplicates())
indep_asso.head(1)

drugbank = pd.read_table("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/drug_set.txt", header=None)
drugbank = drugbank.rename(columns = {0:'DrugBank_ID', 1:'Drug_Name', 2:'InChI', 3:'SMILES', 4:'State'})
drugbank.head(1)

indep_drug_id = indep_asso['drug'].drop_duplicates()
#indep_drug의 순서대로 drugbank를 정렬하기.
df_indep_drug_smiles = drugbank[drugbank['DrugBank_ID'].isin(indep_drug_id)]
sorter = indep_drug_id.tolist()
df_indep_drug_smiles.DrugBank_ID = df_indep_drug_smiles.DrugBank_ID.astype("category")
df_indep_drug_smiles.DrugBank_ID.cat.set_categories(sorter, inplace=True)
df_indep_drug_smiles = df_indep_drug_smiles.sort_values(["DrugBank_ID"])
print len(df_indep_drug_smiles)
df_indep_drug_smiles.head(1) # drug bank id로 중복체크 했을 때 중복 없었음.

new_asso = indep_asso[indep_asso['drug'].isin(df_indep_drug_smiles['DrugBank_ID'])]

new_asso.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/7_new_association/10_PREDICT_association_indep.tsv",sep='\t', index=False)
new_asso.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/10_PREDICT_association_indep.tsv", sindex=False)

import pybel
df_sim = pd.DataFrame()
drug_id = df_indep_drug_smiles['DrugBank_ID'].tolist()
for i in drug_id:
    smi1 = df_indep_drug_smiles[df_indep_drug_smiles.DrugBank_ID == i]['SMILES'].tolist()[0]
    for j in drug_id:
        smi2 = df_indep_drug_smiles[df_indep_drug_smiles.DrugBank_ID == j]['SMILES'].tolist()[0]
        
        smiles = [smi1, smi2]
        mol = [pybel.readstring('smi',x) for x in smiles]
        fps = [x.calcfp(fptype='FP2') for x in mol]
        sim = fps[0] | fps[1]
        
        temp_df = pd.DataFrame([[i, j, sim]], columns = ['DrugBank_ID1','DrugBank_ID2', 'CheSim'])
        df_sim = df_sim.append(temp_df)

df_sim.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/9_indep_drug_similarity_structure.tsv", sep='\t', index=False)
df_sim.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/9_indep_drug_similarity_structure.tsv", sep='\t', index=False)

df_sim = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/9_indep_drug_similarity_structure.tsv")

##### phenotypic similaritied

'''2017.4.21. indep drug X gold drug 간의 phenotype similarity가 필요'''

indep_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/new_predict/3_indep_learning_asso.tsv")
print len(indep_asso['drug_id'].drop_duplicates())
gold_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/new_predict/1_gold_learning_asso.tsv")
print len(gold_asso['drug_id'].drop_duplicates())

indep_se = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/PREDICT_data/3_2_indep_drug_se_matrix.tsv", index_col=0)
gold_se = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/PREDICT_data/4_2_gold_se_by_indep_matrix.tsv", index_col=0)

print len(indep_se)
indep_se.head(1)

print len(gold_se)
gold_se.head(1)

gold_dr_id = gold_asso['drug_id'].drop_duplicates()
indep_dr_id = indep_asso['drug_id'].drop_duplicates()

indep_phe_sim_mat = pd.DataFrame(columns = gold_dr_id.tolist(), index = indep_dr_id.tolist())
indep_phe_sim_mat = indep_phe_sim_mat.fillna(0.0)
print len(indep_phe_sim_mat)
indep_phe_sim_mat.head(1)

from sklearn.metrics import jaccard_similarity_score
for i_dr in indep_dr_id:
    for g_dr in gold_dr_id:
        i_se = indep_se[i_dr]
        g_se = gold_se[g_dr]
        
        jac_sim = jaccard_similarity_score(i_se.tolist(), g_se.tolist())
        
        indep_phe_sim_mat.set_value(i_dr, g_dr, jac_sim)

indep_phe_sim_mat.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/new_predict/2_2_2_indep_drug_gold_drug_se_sim_mat.tsv", sep='\t')

indep_desc = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/8_new_training/2_2_indep_asso_descriptor.tsv")

indep_drug_feature = indep_desc[['drug_id']+indep_desc.columns.values[3:2279].tolist()]
print indep_drug_feature.shape

indep_drug_feature = indep_drug_feature.drop_duplicates('drug_id')
print indep_drug_feature.shape
indep_drug_feature.head(2)

# 아래 for loop에서 index 지정 에러가 떠서 0으로 통일 시킴.
indep_drug_feature.index = [0 for i in range(len(indep_drug_feature))]

from sklearn.metrics import jaccard_similarity_score
df_jac = pd.DataFrame()
drug_id = df_indep_drug_smiles['DrugBank_ID'].tolist()
for i in drug_id:
    desc1 = indep_drug_feature[indep_drug_feature.drug_id == i].loc[0][1:]
    for j in drug_id:
        desc2 = indep_drug_feature[indep_drug_feature.drug_id == j].loc[0][1:]
        
        jac_sim = jaccard_similarity_score(desc1.tolist(),desc2.tolist())
        
        temp_df = pd.DataFrame([[i, j, jac_sim]], columns = ['DrugBank_ID1','DrugBank_ID2', 'PheSim'])
        df_jac = df_jac.append(temp_df)

df_jac.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/9_indep_drug_similarity_phenotype.tsv", sep='\t', index=False)
df_jac.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/9_indep_drug_similarity_phenotype.tsv", sep='\t', index=False)

#### drug similarities 합치는 과정
drug_ch = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/9_indep_drug_similarity_structure.tsv")
drug_ph = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/9_indep_drug_similarity_phenotype.tsv")
df_drug_sim = pd.merge(left = drug_ch, right= drug_ph, left_on = ['DrugBank_ID1', 'DrugBank_ID2'], right_on = ['DrugBank_ID1', 'DrugBank_ID2'])
df_drug_sim = df_drug_sim.rename(columns = {'Similarity_Score_x' : 'CheSim', 'Similarity_Score_y':'PheSim'})
df_drug_sim.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/9_indep_drug_similarity_whole.tsv", index=False, sep='\t')
df_drug_sim.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/9_indep_drug_similarity_whole.tsv", index=False, sep='\t')
