==> Drug_Repositioning/5_new_drug/13_indep2_drug_name_to_drugbank_id.ipynb <==
import pandas as pd
import numpy as np
from datetime import datetime

drdi = pd.read_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/drug_disease_ndf_rt.csv", index_col = 0)
drugbank = pd.read_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/drugbank vocabulary.csv")
#pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/map_umls_2_mesh.tsv")

drug_name = pd.Series(drdi.index.values.tolist())

print len(drug_name)
print len(drugbank)

mapped = drugbank[drugbank['Common name'].isin(drug_name)][['DrugBank ID', 'Common name']]

print len(mapped)
mapped.head(2)

mapped['query_name'] = mapped['Common name']

mapped.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/12_1_drug_name_matching_exact.tsv", sep='\t', index=False)
mapped.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/12_1_drug_name_matching_exact.tsv", sep='\t', index=False)

# NaN 값이 있어서 str.contains 함수를 쓸 수 없어 추가함.
drugbank = drugbank.fillna("None")

query_name = drug_name[~drug_name.isin(drugbank['Common name'])]
print len(query_name)

ex_match = pd.DataFrame()
multi_match = pd.DataFrame()
for i in query_name:
    found = drugbank[drugbank['Synonyms'].str.lower().str.contains(i.lower())]
    if len(found) == 1:
        temp = found[['DrugBank ID', 'Common name', 'Synonyms']]
        temp['query_name'] = i
        ex_match = ex_match.append(temp)
    elif len(found) > 1:
        temp2 = found[['DrugBank ID', 'Common name', 'Synonyms']]
        temp2['query_name'] = i
        multi_match = multi_match.append(temp2)

ex_match.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/12_2_drug_name_mathcing_sym_exact.tsv", sep='\t', index=False)
ex_match.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/12_2_drug_name_mathcing_sym_exact.tsv", sep='\t', index=False)

multi_match.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/12_3_drug_name_mathcing_sym_multi.tsv", sep='\t')
multi_match.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/12_3_drug_name_mathcing_sym_multi.tsv", sep='\t')

# drug name , drugbank id mapping 파일 합치기

ex_match.head(2)

multi_match_man = pd.read_table("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/12_3_2_drug_name_mathcing_sym_multi_mannual.txt")

df1 = mapped[['DrugBank ID', 'query_name']]
df2 = ex_match[['DrugBank ID', 'query_name']]
df3 = multi_match_man[['DrugBank ID', 'query_name']]
frames = [df1, df2, df3]
all_mapped = pd.concat(frames)

print len(all_mapped)
print len(all_mapped.drop_duplicates())

all_mapped.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/12_4_indep2_drugName_to_drugbankID.tsv", sep='\t', index=False)
all_mapped.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/12_4_indep2_drugName_to_drugbankID.tsv", sep='\t', index=False)

# drug 들 중에 side effect가 있는 애들.
sider = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/3_3_SIDER_STITCHid_DrugBankid.tsv")
offside = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/4_3_OFFSIDE_STITCHid_DrugBankid.tsv")

drugs = pd.concat([sider['DrugBank_ID'], offside['DrugBank_ID']])
unique_drugs = drugs.drop_duplicates()

len(unique_drugs)

filtered_drug = all_mapped[all_mapped['DrugBank ID'].isin(unique_drugs)]
print len(filtered_drug)

# side effect가 있으면서, gold standard, independent drug가 아닌 애들
gold = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/7_2_gold_drug_feature_id.tsv")
indep = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/7_3_indep_drug_feature_id.tsv")

len(gold['DrugBank_ID'].drop_duplicates())

len(indep['DrugBank_ID'].drop_duplicates())

used_drugs = pd.concat([gold['DrugBank_ID'].drop_duplicates(),indep['DrugBank_ID'].drop_duplicates()])
print len(used_drugs)
print len(used_drugs.drop_duplicates())

indep[~indep['DrugBank_ID'].isin(gold['DrugBank_ID'].drop_duplicates())].drop_duplicates('DrugBank_ID')

used_drugs[used_drugs.duplicated(keep=False)].sort_values()

일단 보류. --> unique한 association으로 비교해야함.
