==> Drug_Repositioning/5_new_drug/15_PREDICT_indep_similarity_score_0411.ipynb <==
import numpy as np
import pandas as pd
from datetime import datetime

indep_asso = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/PREDICT_data/3_2_indep_learning_asso.tsv")

# drug-drug similarity(structure)

str_sim = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/9_indep_drug_similarity_structure.tsv")
se_sim = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/9_indep_drug_similarity_phenotype.tsv")

indep_dr_id = indep_asso['drug_id'].drop_duplicates().tolist()
indep_di_id = indep_asso['disease_id'].drop_duplicates().tolist()
print len(indep_dr_id)
print len(indep_di_id)

str_dr_id = str_sim['DrugBank_ID1'].drop_duplicates() # structure 정보를 가지고 있는 drug --> 그대로 사용하면됨.
print len(str_dr_id)

str_sim_mat = pd.DataFrame(index = str_dr_id.tolist(),columns=str_dr_id.tolist())
str_sim_mat.fillna(0.0)

for index, row in str_sim.iterrows():
    str_sim_mat[row['DrugBank_ID2']][row['DrugBank_ID1']] = row['CheSim']

str_dr_id.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/new_predict/2_indep_drug_id.tsv",sep='\t', index=False)
str_sim_mat.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/new_predict/2_indep_drug_drug_str_sim_mat.tsv",sep='\t')
str_sim_mat.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/new_predict/2_indep_drug_drug_str_sim_mat.tsv",sep='\t')

# drug-drug similarity(se)

se_sim_mat = pd.DataFrame(index = str_dr_id.tolist(),columns=str_dr_id.tolist())
se_sim_mat.fillna(0.0)

for index, row in se_sim.iterrows():
    se_sim_mat[row['DrugBank_ID2']][row['DrugBank_ID1']] = row['PheSim']

se_sim_mat.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/new_predict/2_indep_drug_drug_se_sim_mat.tsv",sep='\t')
se_sim_mat.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/new_predict/2_indep_drug_drug_se_sim_mat.tsv",sep='\t')

# association

print len(indep_asso)
temp = indep_asso[indep_asso['drug_id'].isin(str_dr_id)]
print len(temp)

pos = temp[temp.association ==1]
fi_disease_id = pos['disease_id'].drop_duplicates()
fi_learning_asso = temp[temp['disease_id'].isin(fi_disease_id)]
print len(fi_learning_asso)

a = fi_learning_asso[fi_learning_asso.association ==1]
b = fi_learning_asso[fi_learning_asso.association ==0]
print len(a)
print len(b)
print len(a['drug_id'].drop_duplicates())
print len(b['drug_id'].drop_duplicates())
print len(a['disease_id'].drop_duplicates())
print len(b['disease_id'].drop_duplicates())

fi_learning_asso.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/7_new_association/new_predict/3_indep_learning_asso.tsv", sep='\t', index=False)
fi_learning_asso.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/7_new_association/new_predict/3_indep_learning_asso.tsv", sep='\t', index=False)

#disease disease similarity

di_feature = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/8_gold_standard_disease_binary_matrix.tsv", index_col = 0)

di_feature.head(1)

str_fi_disease_id = list()
for i in fi_disease_id:
    str_fi_disease_id.append(str(i))

a = di_feature.filter(items = str_fi_disease_id)
print len(a)
a.head(1)
#a.sum() --> 0인 feature 없었음.

di_ge_sim = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/9_indep_disease_similarity_genetic.tsv")
di_phe_sim = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/9_indep_disease_similarity_mimminerc.tsv", index_col=0)

print len(di_ge_sim)
temp = di_ge_sim[(di_ge_sim['Disease_ID1'].isin(fi_disease_id))]
print len(temp)
fi_disease_ge_sim = temp[(temp['Disease_ID2'].isin(fi_disease_id))]
print len(fi_disease_ge_sim)

print len(di_phe_sim)
temp = di_phe_sim.filter(items = str_fi_disease_id)
fi_di_phe_sim = temp.filter(items = fi_disease_id, axis=0)
print len(fi_di_phe_sim)
fi_di_phe_sim.head(1)

fi_disease_id.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/new_predict/2_indep_disease_id.tsv", sep='\t', index=False)
fi_disease_id.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/new_predict/2_gold_disease_id.tsv", sep='\t', index=False)
fi_di_phe_sim.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/new_predict/2_indep_disease_disease_phe_sim_mat.tsv",sep='\t')
fi_di_phe_sim.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/new_predict/2_indep_disease_disease_phe_sim_mat.tsv",sep='\t')
fi_disease_ge_sim.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/new_predict/2_indep_disease_disease_ge_sim.tsv", sep='\t', index=False)
fi_disease_ge_sim.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/new_predict/2_indep_diesease_disease_ger_sim.tsv", sep='\t', index=False)

fi_disease_ge_sim_mat = pd.DataFrame(index = fi_disease_id.tolist(),columns=fi_disease_id.tolist())
fi_disease_ge_sim_mat.fillna(0.0)

for index, row in fi_disease_ge_sim.iterrows():
    fi_disease_ge_sim_mat[row['Disease_ID2']][row['Disease_ID1']] = row['Similarity_Score']

fi_disease_ge_sim_mat.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/6_new_disease/new_predict/2_indep_disease_disease_ge_sim_mat.tsv",sep='\t')
fi_disease_ge_sim_mat.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/6_new_disease/new_predict/2_indep_disease_disease_ge_sim_mat.tsv",sep='\t')
