==> Drug_Repositioning/5_new_drug/16_fingerprint.ipynb <==
import numpy as np
import pandas as pd
import pybel

drug_id = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/new_predict/1_gold_drug_id.tsv", header=None)
indep_dr_id = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/new_predict/2_indep_drug_id.tsv", header=None)
drugbank = pd.read_table("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/drug_set.txt", header=None)
drugbank = drugbank.rename(columns = {0:'DrugBank_ID', 1:'Drug_Name', 2:'InChI', 3:'SMILES', 4:'State'})
drugbank.head(1)

print len(drug_id)
print len(indep_dr_id)

# gold 에 포함되지 않는 indep만의 drug
len(indep_dr_id[indep_dr_id[0].isin(drug_id[0])])

total_dr_id = pd.concat([drug_id, indep_dr_id]).drop_duplicates()
len(total_dr_id)

drug_smiles = drugbank[drugbank['DrugBank_ID'].isin(total_dr_id[0])]

#f = lambda x: 1 if x else 0
fp2_col = ['FP2_'+str(x) for x in range(1,1025)]
fp2_fps = pd.DataFrame(columns=fp2_col, index =total_dr_id[0].tolist())
fp2_fps = fp2_fps.fillna(0)
fp2_fps['SMILES'] = ''
for index, row in drug_smiles.iterrows():
    mol = pybel.readstring('smi', row['SMILES'])
    fps = mol.calcfp(fptype='FP2')
    for idx in fps.bits:
        fp2_fps.set_value(row['DrugBank_ID'], 'FP2_'+str(idx), 1)
        fp2_fps.set_value(row['DrugBank_ID'], 'SMILES', row['SMILES'])

fp2_fps.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/new_predict/3_drug_fp2_fps.tsv", sep='\t')
fp2_fps.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/new_predict/3_drug_fp2_fps.tsv", sep='\t')

#f = lambda x: 1 if x else 0
maccs_col = ['MACCS_'+str(x) for x in range(1,167)]
maccs_fps = pd.DataFrame(columns=maccs_col, index =total_dr_id[0].tolist())
maccs_fps = maccs_fps.fillna(0)
maccs_fps['SMILES'] = ''
for index, row in drug_smiles.iterrows():
    mol = pybel.readstring('smi', row['SMILES'])
    fps = mol.calcfp(fptype='MACCS')
    for idx in fps.bits:
        maccs_fps.set_value(row['DrugBank_ID'], 'MACCS_'+str(idx), 1)
        maccs_fps.set_value(row['DrugBank_ID'], 'SMILES', row['SMILES'])

maccs_fps.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/new_predict/3_drug_maccs_fps.tsv", sep='\t')
maccs_fps.to_csv("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/new_predict/3_drug_maccs_fps.tsv", sep='\t')

maccs_fps
