==> Drug_Repositioning/5_new_drug/18_drdrSim(matlab,R)_tidy.ipynb <==
# R에서 작업한 drdrGo, matlab에서 작업한 drdrtargetSim을 정리하는 작업.
# 1. header의 순서를 다른 drdr sim mat의 순서와 동일하게 한다.
# 2. gold와 indep 버전을 만든다.

import pandas as pd
import numpy as np

targetSim = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/new_predict/1_drug_drug_target_sim_mat.tsv", index_col=0)
strSim = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/new_predict/1_drug_drug_str_sim_mat.tsv", index_col=0)
goSim_max = pd.read_table("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/new_predict/drug_gosemsim_resnik_max.txt", index_col=0)
goSim_bma = pd.read_table("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/new_predict/drug_gosemsim_resnik_BMA.txt", index_col=0)

print targetSim.isnull().any().any(), goSim_max.isnull().any().any(), goSim_bma.isnull().any().any()
targetSim.fillna(0.0, inplace=True)
goSim_max.fillna(0.0, inplace=True)
goSim_bma.fillna(0.0, inplace=True)
print targetSim.isnull().any().any(), goSim_max.isnull().any().any(), goSim_bma.isnull().any().any()

# column순서를 기존의 drdr sim mat순서와 맞춘다.
order = strSim.columns.values
targetSim = targetSim[order.tolist()]
goSim_max = goSim_max[order.tolist()]
goSim_bma = goSim_bma[order.tolist()]

print order.tolist() == targetSim.columns.tolist(),order.tolist() == goSim_max.columns.tolist(),order.tolist() == goSim_bma.columns.tolist()

# index 순서를 기존의 drdr sim mat순서와 맞춘다.
targetSim = targetSim.reindex(index = strSim.index)
goSim_max = goSim_max.reindex(index = strSim.index)
goSim_bma = goSim_bma.reindex(index = strSim.index)

idx=strSim.index.tolist()
print idx == targetSim.columns.tolist(),idx == goSim_max.columns.tolist(),idx == goSim_bma.columns.tolist()

targetSim.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/new_predict/1_drug_drug_targetSim_mat.tsv", sep='\t')
goSim_max.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/new_predict/1_drug_drug_GoSim_max_mat.tsv", sep='\t')
goSim_bma.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/new_predict/1_drug_drug_GoSim_bma_mat.tsv", sep='\t')

# 2.indep 버전을 만든다. (target은 이미 있으므로 drGo bma만. target의 순서가 같은지는 모르겠음. but 사실 순서는 큰 상관이 없음.)

indep_strSim = pd.read_table("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/new_predict/2_2_2_indep_drug_gold_drug_str_sim_mat.tsv", index_col = 0)

indep_drGo_bma = pd.read_table("/home/share/aschoi/nas/users/asolchoi/data/Drug_Repositioning/5_new_drug/new_predict/indep_dr_gosemsim_resnik_BMA.txt", index_col=0)

print indep_strSim.shape
print indep_drGo_bma.shape
indep_order = indep_strSim.columns.values

indep_goSim = indep_drGo_bma.filter(items = indep_strSim.index, axis=0)
indep_goSim = indep_goSim[indep_strSim.columns.tolist()]
indep_goSim = indep_goSim.reindex(index = indep_strSim.index)
print indep_strSim.columns.tolist() == indep_goSim.columns.tolist()
print indep_strSim.index.tolist() == indep_goSim.index.tolist()

indep_goSim.fillna(0.0, inplace=True)
indep_goSim.isnull().any().any()

indep_goSim.to_csv("/DAS_Storage1/aschoi/data/Drug_Repositioning/5_new_drug/new_predict/2_2_2_indep_drug_gold_drug_GoSim_bma_mat.tsv", sep='\t')
